Refactored crc function
